Adding file changes and Minor Refactor
Added missing file changes, and did a minor refactor on Error
